requirements.txt
annotated-types==0.6.0
arrow==1.2.3
asgiref==3.7.2
beautifulsoup4==4.12.2
blessed==1.20.0
blis==0.7.11
bs4==0.0.1
cairocffi==1.6.1
CairoSVG==2.7.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.3
cssselect2==0.7.0
cymem==2.0.8
defusedxml==0.7.1
Django==4.2.3
django-picklefield==3.1
django-q==1.3.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.0
polygon-api-client==1.13.3
preshed==3.0.9
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
pytz==2023.3.post1
redis==3.5.3
regex==2023.8.8
reportlab==4.0.7
requests==2.31.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
svglib==1.5.1
thinc==8.2.1
tinycss2==1.2.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
vaderSentiment==3.3.2
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.3
webencodings==0.5.1
websockets==12.0
